adicionado o react navigator
