+++ screens/Home/index.js
@@ -1,0 +1,22 @@
+
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+const Home = () => {
+    return (
+        <View style={styles.container}>
+        <Text>Esta é a home!!!</Text>
+        </View>
+
+    );
+}
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      justifyContent:'center',
+      alignItems:'center'
+    },
+  });
+
+export default Home;
